refactor: #20 解答フォーム削除ボタンを配置するためAnswerAreaを作成
